Add multidimensional urlencode
Go depends on arguments being in the `some_object[key]=value` format,
which is fairly common in web development, but this isn't part of the core
urlencode. Luckily uber has made a version available that does support that.

But since one of my goals is for this library to be packaged up on its own
without any external dependencies I've ended up having to vendor it.