[#2] Encode POST data when not None
The underlying urllib library reacts weirdly when non-string values
is passed through on certain versions of Python. For some reason
passing a dict worked in Python 2.6 but not 2.7.

From now on the definition of a something that should be a POST
request is:

  - Any string (even empty)
  - Any dict (even empty)
  - True (which converts into an empty string)

Note that list is not part of this as I don't know what a sensible
way to handle them would be. If the scenario comes up I'll deal with
it then. I imagine possibly key/value pairs, but I don't see that as
a scenario that would be preferable to a dict at the moment.