Add automatic retry to Artifact.get_directory
Since Go returns a 202 by default it shouldn't be up to the standard
user to have to keep track of having to retry.

The values used for backoff and so on are pretty random and have
worked well enough for me in testing. Hoping it'll be good enough for
most use cases.