Make debug output from httplib work
Previously it wasn't being set properly as the value was being set at
init, despite there being an attribute for it on the server class.

By setting the debug flag on init and then ensuring that the urllib
opener is always being set (with or without the auth handler) the debug
output will actually be shown.

This relates to pr gaqzi/gocd/#11 and issue gaqzi/gocd-cli/#14.